requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
motor==3.7.1           # MongoDB async driver
python-dotenv==1.1.1   # Load environment variables
passlib[bcrypt]==1.7.4 # Password hashing
python-jose[cryptography]==3.5.0 # JWT authentication
pydantic[email]==2.11.7
python-multipart==0.0.20 # File uploads
email-validator==2.1.0.post1
httpx==0.28.1          # Async HTTP requests (optional)
pytest==7.4.2           # Testing
